requirements.txt
appdirs==1.4.3
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
lxml==4.1.0
MySQL-python==1.2.5
pytils==0.3
pytz==2017.2
six==1.11.0
